Created files for the start up page
Created three basic files for the start up page, the HTML, CSS and
JavaScript file. I don't intend to do much about the HTML and
JavaScript, I intend to focus on studying how to use CSS

Started with the boilerplate code for the html file. The CSS and
JavaScript File are still blank.