Practiced with Copying HTML Element Styles by Eye
The video tutorial I am following, have an exercise to get used
with the basics of CSS and HTML, I have prior knowledge of the
basics of HTML Elements.

The exercise involves styling an HTML button using CSS, setting up the
size, color, and position of the button element with regards to
contents.