Practice styling button element with CSS
Practiced setting up the Color, position, and conditional behaviour of
an html element.

Trying out the use of pseudo-classes for elements. The Hover and Active
pseudo-classes are useful for showing the

Usage of padding and margin over setting values to the height and width
property of an element. Margin is the distance of an element over other
elements, and the Padding is the distance of the content of the element
over the border or the edge of an element.

Margin affects the actual position of an html element in the webpage,
which may cause unexpected behaviours with the hover style of an
element. I caused an error wherein upon hovering on an element,the
margin would drstically change, upon activating the hover the style
immediately applies, which changes the margin then the position of the
hovered element. Upon activation the style is applied and de-applied
to the element, the element looks like it is in constantly and rapidly
changing position when hovered.

I got to try out the transition property, and I don't understand the
many arguments when using setting up the value of the property.

A quick peek of the CSS box model, the purpose of the CSS Box model,
how it affects the styling of an element.

The opacity of an element can be modified through CSS through the
background-color property, I cannot understand why there are so many
varieties of ways to set the color of an element. The a in rgba stands
for the opacity of an element.