Practiced styling text with CSS
I wanted to try practicing how to modify the look of a text with CSS,
I tried using pseudo-classes, and the only pseudo-classes I know
of is hover and active, I used hover mostly.

Changed the font-weight and background-color property because I find
this property the easiest to work with in order to see a quick
change in terms of how a text looks and feels visually.

I wanted to practice styling elements with textual content, so
I used general html elements in order to group a bunch of semantic
html elements.

Combining div and span in order to group and single certain elements,
I used span in order to single-out certain parts of the p element.

I wanted to select a portion of a text to have a specific and
unique color and font size from the rest of the other parts of
the paragraph.

So I used the span element within the p html element, I am not sure
of where can I use the span element outside of the paragraph html
element.

I coded along with a CSS tutorial, and tried out the try-it-yourself
practices in order to make practical use of what I watched, though in
terms of organizing the outputs from the practice, I find the
final result cluttered and unorganized because i don't have a plan
for the long-term maintenance of the code.

And I did not think much of who will ever have to read the code, so
the objective was to put something together fast and put it outside
and call it quits.

NOTE:
I forgot most of what I was thinking when I was coding

NOTE:
I am not sure of which or what I was doing when I was writing the
code, because as of writing this commit message, two days have passed
since I actually wrote the code, and with all that has happened,
I lost track of what I am doing because I did not code yesterday,
because of circumstances, did not have extension wire.