Practice using display property with HTML elements
The exercise involves copying the design of a youtube page showing
multiple elements. With the utilization of diplay property to change
the way how elements are displayed.

Utilization of display property and div element ( wrapper element for
HTML elements )

Created p and img elements to start with position modification test
and experiments.