Created files to practice designing div elements
I want to try modifying the property of div elements to observe how
the contents of a div is displayed and behave according to the set
values of the display property.

There are default values, the default values can be changed through
the linked CSS file, inline-styling, or through a style block within
the head.

Added new styles to videoHolder element to fix the spacing between
the div and outside elements.S