finished practice with NESTED LAYOUT
I tried out combining horizontal and vertical layout to copy the
design of youtube.

I learned that inline-block can allow for multiple div elements
be presented in a line. There are problems when the width of the
elements is not satisfied by the available space, which the element
proceeds to take another line.

Created a folder to contain images that will be used as the channels
icon. The images to be used as channel icons is from youtube, I used
the built-in browser developer tools  (inspect ) to retrieve
the images from the website ( youtube ).

I copied word for word how the tutorial managed resources, I
find that I suck in resource management, and often I Create
disorganized mess when making large-scale projects. This time I want
to follow through the tutorial word for word and understand the
reasons behind design and structural desicions.