Renamed the folder names, no further changes.
I just renamed the folders. Then the version control extension in
visual studio code shows that the changes are inside of all of the
files.

I don't know why all of the files inside of the folders are now
considered to be new additions.

I wanted to rename the folders to relect their contents, since I find
it confusing, when I created the folders, their meaning and purpose
did seem clear, but now that I am making some progress on new files and
left this folder untouched in preparation for finals.

I forgot what
the original files are supposed to do and why exactly did I create them.