Renamed the HTML file for Exercise Five
I renamed the HTML file for exercise five because the name is the same
as the name for the HTML file for exercise Three.

1.Name is not descriptive Enough.
2.Name is SIMILAR TO ANOTHER FILE, I'm confused.

After the rename, git considers it as a new file and shows the previous
file to be crossed out. I think git treats the renamed file different
from the original file. Which makes sense because the names are
different, in terms of content, I didn't do any changes after the
rename.

Though I made sure that all of the code changes were finished first
before I renamed the file.  I did not commit the changes first, and
proceeded to continue working on another file.

NOTE:
I think I should have commited the changes for the code first
before I did any renaming operations because I cannot see any changes
highlighted in the editor. Everything is highlighted. I cannot see
which were changed and which weren't changed.