Copied ExerciseFive Folder for base to Exercise 6
Copied the folder for ExerciseFive to be used as the base for the
next exercise, exercise Six involves using multiple stylesheets to
style different sections of a website.

Renamed the BASE HTML, CSS, and JAVASCRIPT files to be used as starting
point for the exercise.

Through this exercise, I want to practice using multiple CSS files for
specialized purposes.