Started progress with using Flex for site Header
I want to use Flex in order to style the display of the header for the
website.

Created a separate stylesheet only for the style of the Header Element,
I'm calling it header because it is at the top of all the elements
in the website. The header contains the search bar and other icons
for navigating the content of the website.

Added more dummy videoHolder in the HTML file, I want to mimic the
display youtube has with the video content available on the website.