Trying out the functionality of Flex Display
I'm not familiar with how Flex presents elements inside of a div.
I want to practice with Flex and get the feel for using it first,
before I proceed to using flex display in ExerciseSix.

I will probably place it inside of a separate folder. So the name and
history of the files might change.