Trying out Flex display, Made separate stylesheets
Created the videoHolderFlexDisplay.css to contain the styles for the
flex display of the videoHolder.

Copied the exerciseSixStyle.css,
the only modified part is that of the style for the videoHolder
( div element ) and the mainContent ( div element containing the
videoHolderContainer element)

I want the videoHolder (div element) to respond to the changes in
available space in the browser window, should the browser window be
resized, then the image of the video should resize as well.

MY REASON FOR USING DUAL STYLESHEETS:
1. I used two seperate stylesheets because I want to practice
working with more than one stylsheets that applies to a website.

2. I want to separate the original stylesheet and create a stylesheet
to handle all the experimentations and test.

In doing so, I want to acheive the horizontal space adaptability of
flex and the column consistency of grid.

I'm still working on applying it on exerciseSix file.