Starting recreation of search bar & icons from YT
In the video tutorial that I am following, one of the exercises
involved using the flex and grid display to achieve the look of the
top most portion of the YouTube website that holds the search bar
and other icons for modifying the settings.

To begin with, I am copying the technique the tutorial shows to
represent the visual elements of a website.